requirements.txt
Flask
requests
python-dotenv
openai
pathlib
ffmpeg-python